Restructured SAGE directories.
Have split off the functions for R-SAGE into their own directory. These
are split into ``grid`` files and ``selfcon`` files that control
``ReionizationOn`` = 2 and 3 respectively.
