+++ sage/Makefile
@@ -16,12 +16,18 @@ OBJS   = 	./main.o \
 			./model_reincorporation.o \
 			./model_mergers.o \
 			./model_misc.o \
-			./io/tree_binary.o
+			./io/tree_binary.o \
+			./self_consistent/selfcon.o \
+			./self_consistent/grid.o \
+			./self_consistent/common.o 
 
 INCL   =	./core_allvars.h  \
 			./core_proto.h  \
 			./core_simulation.h  \
 			./io/tree_binary.h \
+			./self_consistent/selfcon.h \
+			./self_consistent/grid.h \
+			./self_consistent/common.h \
 			./Makefile
 
 #USE-MPI = yes# set this if you want to run in parallel
