+++ sage/core_save.c
@@ -88,7 +88,7 @@ void save_galaxies(int filenr, int tree)
 
       if(HaloGal[i].SnapNum == max_snap && Halo[HaloGal[i].HaloNr].Descendant == -1)
       {
-          write_gridarray(&HaloGal[i], save_fd[n]); // Input snapshots are ordered highest -> lowest so it'll be 0th element. 
+          write_temporal_arrays(&HaloGal[i], save_fd[n]); // Input snapshots are ordered highest -> lowest so it'll be 0th element. 
           TotGalaxies[n]++;
           TreeNgals[n][tree]++;          
       } 
@@ -98,79 +98,6 @@ void save_galaxies(int filenr, int tree)
 
   // don't forget to free the workspace.
   free( OutputGalOrder );
-
-}
-
-void write_gridarray(struct GALAXY *g, FILE *fp)
-{
-
-#define WRITE_GRID_PROPERTY(name, length)    \
-{                                            \
-  XASSERT(name != NULL, #name" has a NULL pointer.\n"); \
-  nwritten = fwrite(name, sizeof(*(name)), length, fp); \
-  XASSERT(nwritten == length, "While writing "#name", we expected to write %d times but wrote %d times instead\n", \
-          length, nwritten);                 \
-}
-
-#define WRITE_CONVERTED_GRID_PROPERTY(name, length, conversion, type)    \
-{                                            \
-  XASSERT(name != NULL, #name" has a NULL pointer.\n"); \
-  buffer = calloc(length, sizeof(type));\
-  XASSERT(buffer != NULL, "Could not allocate memory for a buffer in write_gridarray for "#name".\n"); \
-  for (j = 0; j < length; ++j)\
-  {\
-    ((type*)buffer)[j] = name[j] * conversion;\
-  }\
-  nwritten = fwrite(buffer, sizeof(type), length, fp); \
-  XASSERT(nwritten == length, "While writing"#name", we expected to write %d times but wrote %d times instead\n", \
-          length, nwritten);                 \
-  free(buffer); \
-}
-
-  float SFR_conversion = UnitMass_in_g / UnitTime_in_s * SEC_PER_YEAR / SOLAR_MASS / STEPS; 
-  int j;  
-  int32_t nwritten;
-  void *buffer;
- 
-  nwritten = fwrite(&g->TreeNr, sizeof(g->TreeNr), 1, fp);
-
-  fwrite(&g->mergeType, sizeof(g->mergeType), 1, fp);
- 
-  XASSERT(g->IsMalloced == 1, "We are trying to write out the grid arrays for a galaxies who has already been freed.\n");
- 
-
-  WRITE_GRID_PROPERTY(g->GridHistory, MAXSNAPS);
-  WRITE_GRID_PROPERTY(g->GridStellarMass, MAXSNAPS);
-
-  WRITE_CONVERTED_GRID_PROPERTY(g->GridSFR, MAXSNAPS, SFR_conversion, typeof(*(g->GridSFR))); 
-
-  WRITE_GRID_PROPERTY(g->GridZ, MAXSNAPS);
-  WRITE_GRID_PROPERTY(g->GridFoFMass, MAXSNAPS);
-  WRITE_GRID_PROPERTY(g->EjectedFraction, MAXSNAPS);
-  WRITE_GRID_PROPERTY(g->LenHistory, MAXSNAPS);
-
-  WRITE_CONVERTED_GRID_PROPERTY(g->GridOutflowRate, MAXSNAPS, SFR_conversion, typeof(*(g->GridOutflowRate)));
-  WRITE_CONVERTED_GRID_PROPERTY(g->GridInfallRate, MAXSNAPS, SFR_conversion, typeof(*(g->GridInfallRate)));
-
-  WRITE_GRID_PROPERTY(g->GridEjectedMass, MAXSNAPS);
-  WRITE_GRID_PROPERTY(g->QuasarActivity, MAXSNAPS);
- 
-  WRITE_CONVERTED_GRID_PROPERTY(g->DynamicalTime, MAXSNAPS, UnitTime_in_Megayears, typeof(*(g->DynamicalTime)));
-
-  WRITE_GRID_PROPERTY(g->QuasarSubstep, MAXSNAPS);
-  WRITE_GRID_PROPERTY(g->GridColdGas, MAXSNAPS);
-  WRITE_GRID_PROPERTY(g->LenMergerGal, MAXSNAPS);
-  WRITE_GRID_PROPERTY(g->GridBHMass, MAXSNAPS);
-  WRITE_GRID_PROPERTY(g->GridReionMod, MAXSNAPS);
-  WRITE_GRID_PROPERTY(g->GridDustColdGas, MAXSNAPS);
-  WRITE_GRID_PROPERTY(g->GridDustHotGas, MAXSNAPS);
-  WRITE_GRID_PROPERTY(g->GridDustEjectedMass, MAXSNAPS);
-  WRITE_GRID_PROPERTY(g->GridType, MAXSNAPS);
-
-#undef WRITE_GRID_PROPERTY
-#undef WRITE_CONVERTED_GRID_PROPERTY
- 
-  ++times_written;
 
 }
 
@@ -238,7 +165,7 @@ void save_merged_galaxies(int filenr, int tree)
 
   for(i = 0; i < MergedNr; i++)
   {
-    write_gridarray(&MergedGal[i], save_fd2);
+    write_temporal_arrays(&MergedGal[i], save_fd2);
 
     TotMerged++;
     TreeNMergedgals[tree]++;
