+++ sage/self_consistent/common.h
@@ -1,0 +1,7 @@
+#ifndef COMMON_H 
+#define COMMON_H 
+
+// Proto-Types //
+
+
+#endif
