+++ sage/self_consistent/grid.h
@@ -1,0 +1,6 @@
+#ifndef SELFCON_H
+#define SELFCON_H
+
+// Proto-Types //
+
+#endif
