+++ sage/self_consistent/selfcon.h
@@ -1,0 +1,8 @@
+#ifndef SELFCON_H
+#define SELFCON_H
+
+// Proto-Types //
+
+double search_for_modifier(int64_t match_HaloID, int32_t SnapNum, int32_t infall, int64_t *found_idx);
+
+#endif
